API * Refactor Hero to implement and interface of IHero.
* `evolve` method increment all stats of hero with a *multiply of half the original stat vavlue.
