+++ InterviewTest/Interfaces/IHero.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace InterviewTest.Controllers
+namespace InterviewTest.Interfaces
 {
     public interface IHero
     {
